Added Psych to the Gemfile, and updated Pharmacy seeds to use it. Resolves #20.
Psych has an interface change vs. Syck, which causes #20. Adding Psych to the
Gemfile forces Rails to use Psych as the YAML parser, so that everyone will be
using the same parser (instead of it being config dependent). A matching change
to the seeds file has been added.

Note that if you're using the VM, you'll have to apt-get install libyaml-dev
before rerunning bundle install.